requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
pydantic==2.5.3
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
aiohttp-socks==0.8.4